+++ pkg/services/watch_endpoints.go
@@ -10,6 +10,7 @@ import (
 	"github.com/kube-vip/kube-vip/pkg/debouncer"
 	"github.com/kube-vip/kube-vip/pkg/endpoints"
 	"github.com/kube-vip/kube-vip/pkg/endpoints/providers"
+	"github.com/kube-vip/kube-vip/pkg/metrics"
 	"github.com/kube-vip/kube-vip/pkg/servicecontext"
 	"github.com/kube-vip/kube-vip/pkg/utils"
 	v1 "k8s.io/api/core/v1"
@@ -18,6 +19,10 @@ import (
 
 func (p *Processor) watchEndpoint(svcCtx *servicecontext.Context, id string, service *v1.Service,
 	provider providers.Provider, cancelWatcher context.CancelCauseFunc) error {
+	watcherLoops := metrics.WatcherLoops.WithLabelValues("endpoint")
+	watcherLoops.Inc()
+	defer watcherLoops.Dec()
+
 	log.Info("watching", "provider", provider.GetLabel(), "service_name", service.Name, "namespace", service.Namespace)
 	// Use a restartable watcher, as this should help in the event of etcd or timeout issues
 
@@ -83,8 +88,14 @@ func (p *Processor) watchEndpoint(svcCtx *servicecontext.Context, id string, ser
 				log.Info("[endpoint watcher] endpoint object deleted", "provider", provider.GetLabel(), "service name", service.Name, "namespace", service.Namespace)
 			}
 
-			restart, err := epProcessor.Reconcile(svcCtx, event, &lastKnownGoodEndpoint, service, id,
-				p.StartServicesLeaderElection, &wg, p.clientSet, p.updateEgressConfiguration)
+			var restart bool
+			var err error
+			if !p.withActiveService(svcCtx, func() {
+				restart, err = epProcessor.Reconcile(svcCtx, event, &lastKnownGoodEndpoint, service, id,
+					p.StartServicesLeaderElection, &wg, p.clientSet, p.updateEgressConfiguration)
+			}) {
+				return nil
+			}
 			if restart {
 				continue
 			} else if err != nil {
@@ -95,6 +106,7 @@ func (p *Processor) watchEndpoint(svcCtx *servicecontext.Context, id string, ser
 			if svcCtx.Ctx.Err() != nil {
 				return nil
 			}
+			metrics.WatcherFailuresTotal.WithLabelValues("endpoint", "watch_error").Inc()
 			watchErr := utils.WatchError(event.Object)
 			log.Error("watch error", "provider", provider.GetLabel(), "err", watchErr)
 			return utils.WrapPanicError(watchErr, "[%s] endpoint watch failed", provider.GetLabel())
@@ -103,6 +115,7 @@ func (p *Processor) watchEndpoint(svcCtx *servicecontext.Context, id string, ser
 	if svcCtx.Ctx.Err() != nil {
 		return nil
 	}
+	metrics.WatcherFailuresTotal.WithLabelValues("endpoint", "channel_closed").Inc()
 	log.Info("[endpoint watcher] stopping watching", "provider", provider.GetLabel(), "service name", service.Name, "namespace", service.Namespace)
 	return utils.NewPanicError("[%s] endpoint watch channel closed unexpectedly for service %s/%s", provider.GetLabel(), service.Namespace, service.Name)
 }